+++ kibot/gs.py
@@ -115,6 +115,8 @@ class GS(object):
     kikit_units_to_kicad = {'mm': IU_PER_MM, 'cm': 10*IU_PER_MM, 'dm': 100*IU_PER_MM,
                             'm': 1000*IU_PER_MM, 'mil': IU_PER_MILS, 'inch': 1000*IU_PER_MILS,
                             'in': 1000*IU_PER_MILS}
+    # Maximum recursive replace
+    MAXDEPTH = 20
     #
     # Global defaults
     #